grammar check up to ...
The dominant coupling due to the power entering the SRC can be intuitively understood by considering a Fabry-Perot cavity. For small changes in cavity length, the light transmitted by a cavity that is on resonance or anti-resonance will be modulated by a term proportional to the square of the length change. This signal due to the length of the SRC is noise. The amplitude of this noise is linearly dependent on the amount of light entering the cavity which is a result of the DC offset. By decreasing the DC offset, this noise is decreased. Unlike the gravitational wave signal, as this noise is in the amplitude quadrature of the signal beam it will not be amplified by the LO.